# docs/openapi.yaml
openapi: 3.0.0
info:
  title: ARP Guard API
  description: API for ARP Guard - Network Security and Device Management
  version: 1.0.0
  contact:
    name: ARP Guard Team
    email: team@arpguard.io
    url: https://arpguard.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.arpguard.io
    description: Production server
  - url: https://api.arpguard.io/v1
    description: Production server (explicit version)
  - url: https://staging.arpguard.io/api
    description: Staging server
...
paths:
...
  /stats:
    get:
      summary: Get system statistics
      description: Get statistics about the system
      tags:
        - System
      security:
        - apiKey: []
      responses:
        '200':
          description: System statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stats'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /alerts:
    get:
      summary: List all alerts
      description: Get a list of all alerts
      tags:
        - Alerts
      security:
        - apiKey: []
      parameters:
        - name: status
          in: query
          description: Filter alerts by status
          schema:
            type: string
            enum: [open, acknowledged, resolved]
        - name: severity
          in: query
          description: Filter alerts by severity
          schema:
            type: string
            enum: [low, medium, high, critical]
        - name: start_date
          in: query
          description: Filter alerts by start date (ISO format)
          schema:
            type: string
            format: date-time
        - name: end_date
          in: query
          description: Filter alerts by end date (ISO format)
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            default: 50
            maximum: 100
      responses:
        '200':
          description: List of alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Alert'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
    apiKeySecret:
      type: apiKey
      in: header
      name: X-API-Secret
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    RateLimited:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          description: Rate limit ceiling for the given endpoint
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: Number of requests left for the time window
          schema:
            type: integer
        X-RateLimit-Reset:
          description: Time at which the rate limit resets in UTC epoch seconds
          schema:
            type: integer
        X-RateLimit-Retry-After:
          description: Seconds to wait before retrying the request
          schema:
            type: integer
        X-RateLimit-Type:
          description: Type of rate limit (IP, user, token, or endpoint)
          schema:
            type: string
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  retry_after:
                    type: integer
                    description: Seconds to wait before retrying the request
  schemas:
...
    Alert:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the alert
        device_id:
          type: string
          description: ID of the device related to the alert
        severity:
          type: string
          description: Severity level of the alert
          enum: [low, medium, high, critical]
        status:
          type: string
          description: Current status of the alert
          enum: [open, acknowledged, resolved]
        message:
          type: string
          description: Alert message
        details:
          type: object
          description: Additional details about the alert
        timestamp:
          type: string
          format: date-time
          description: Timestamp when the alert was generated
        acknowledged_at:
          type: string
          format: date-time
          nullable: true
          description: Timestamp when the alert was acknowledged
        resolved_at:
          type: string
          format: date-time
          nullable: true
          description: Timestamp when the alert was resolved
      required:
        - id
        - device_id
        - severity
        - status
        - message
        - timestamp
...
    Stats:
      type: object
      properties:
        devices:
          type: object
          properties:
            total:
              type: integer
              description: Total number of devices
            active:
              type: integer
              description: Number of active devices
            inactive:
              type: integer
              description: Number of inactive devices
        alerts:
          type: object
          properties:
            total:
              type: integer
              description: Total number of alerts
            open:
              type: integer
              description: Number of open alerts
            acknowledged:
              type: integer
              description: Number of acknowledged alerts
            resolved:
              type: integer
              description: Number of resolved alerts
            by_severity:
              type: object
              properties:
                low:
                  type: integer
                  description: Number of low severity alerts
                medium:
                  type: integer
                  description: Number of medium severity alerts
                high:
                  type: integer
                  description: Number of high severity alerts
                critical:
                  type: integer
                  description: Number of critical severity alerts
      required:
        - devices
        - alerts
...
    Error:
      type: object
      properties:
        status:
          type: integer
          description: HTTP status code
        code:
          type: string
          description: Error code
        message:
          type: string
          description: Error message
      required:
        - status
        - message 